* Switch the build system to Makeconf
  * Add support for cross-compiling for Android




git-svn-id: svn://svn.code.sf.net/p/libkqueue/code/branches/stable@556 fb4e3144-bc1c-4b72-a658-5bcd248dd7f7

+++ src/linux/vnode.c
@@ -105,7 +105,7 @@ get_one_event(struct inotify_event *dst, int pfd)
             break;
         }
     }
-    dbg_printf("read(2) from inotify wd: %zu bytes", n);
+    dbg_printf("read(2) from inotify wd: %lu bytes", (long) n);
 
     /* FIXME-TODO: if len > 0, read(len) */
     if (dst->len != 0)
